Finalizado função para criar contas e tipo contas padrão. Movido função criarContasPadrao e criarUsuarioAdminInicial para arquivo padrao.go, pertencente ao pacote main
